# docs/swagger.yaml
basePath: /
definitions:
  controller.CreateCategoryRequest:
    properties:
      name:
        type: string
      owner_id:
        type: integer
    required:
    - name
    - owner_id
    type: object
  controller.CreateItemRequest:
    properties:
      category_id:
        type: integer
      condition:
        type: string
      name:
        type: string
      owner_id:
        type: integer
      serial_no:
        type: string
      status:
        type: string
    required:
    - category_id
    - name
    - owner_id
    type: object
  controller.CreateLoanRequest:
    properties:
      borrower_id:
        type: integer
      item_id:
        type: integer
    required:
    - borrower_id
    - item_id
    type: object
  controller.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controller.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  controller.UpdateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  controller.UpdateItemRequest:
    properties:
      condition:
        type: string
      name:
        type: string
      status:
        type: string
    required:
    - condition
    - name
    - status
    type: object
  controller.UpdateLoanRequest:
    properties:
      returned_at:
        type: string
    type: object
  controller.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  helper.PaginatedWebResponse:
    properties:
      code:
        type: integer
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      status:
        type: string
      total:
        type: integer
    type: object
  helper.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
  model.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.Item:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      condition:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      owner:
        $ref: '#/definitions/model.User'
      owner_id:
        type: integer
      serial_no:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.Loan:
    properties:
      borrowed_at:
        type: string
      borrower:
        $ref: '#/definitions/model.User'
      borrower_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      returned_at:
        type: string
      updated_at:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: REST API untuk sistem peminjaman barang P2P
  title: Loans Item API
  version: "1.0"
paths:
  /api/categories:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.PaginatedWebResponse'
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Create category
      tags:
      - categories
  /api/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Delete category
      tags:
      - categories
    get:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Update category
      tags:
      - categories
  /api/items:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.PaginatedWebResponse'
      summary: Get all items
      tags:
      - items
    post:
      consumes:
      - application/json
      parameters:
      - description: Item data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Create item
      tags:
      - items
  /api/items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Delete item
      tags:
      - items
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Item'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Get item by ID
      tags:
      - items
    put:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Update item
      tags:
      - items
  /api/loans:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.PaginatedWebResponse'
      summary: Get all loans
      tags:
      - loans
    post:
      consumes:
      - application/json
      parameters:
      - description: Loan data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateLoanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Create loan (borrow item)
      tags:
      - loans
  /api/loans/{id}:
    delete:
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Delete loan
      tags:
      - loans
    get:
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Loan'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Get loan by ID
      tags:
      - loans
    put:
      consumes:
      - application/json
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Returned at (optional, defaults to now)
        in: body
        name: body
        schema:
          $ref: '#/definitions/controller.UpdateLoanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Return loan (update returned_at)
      tags:
      - loans
  /api/loans/borrower/{borrower_id}:
    get:
      parameters:
      - description: Borrower ID
        in: path
        name: borrower_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.PaginatedWebResponse'
      summary: Get loans by borrower ID
      tags:
      - loans
  /api/users:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.PaginatedWebResponse'
      summary: Get all users
      tags:
      - users
  /api/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Update user
      tags:
      - users
  /api/users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Login user
      tags:
      - users
  /api/users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.WebResponse'
      summary: Register user
      tags:
      - users
swagger: "2.0"